apps/monolith/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
redis>=5.0
confluent-kafka>=2.3
gunicorn>=21.2
python-dotenv>=1.0
dj-database-url>=2.1
django-cors-headers>=4.3.0
stripe>=7.0.0
